pyhton-common: move some tests from cephadm/test_spec.py
Cause they don't have any dependencies to cephadm

(cherry picked from commit 6f90f0fa2efc830eec07373b94534e3c96f24945)
